# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAVVORA E-Commerce API",
    "description": "SAVVORA Production Enterprise E-Commerce Platform API",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/orders/checkout": {
      "post": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Checkout Order",
        "description": "Creates a new order from user's shopping cart with stock reservation and Razorpay payload.",
        "operationId": "checkout_order_api_v1_orders_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/orders": {
      "get": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Get User Orders",
        "description": "Lists order history for logged-in customer.",
        "operationId": "get_user_orders_api_v1_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderOut"
                  },
                  "type": "array",
                  "title": "Response Get User Orders Api V1 Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Get Order Details",
        "description": "Fetches full details of a specific order.",
        "operationId": "get_order_details_api_v1_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Cancel Order",
        "description": "Cancels order and restores stock inventory.",
        "operationId": "cancel_order_api_v1_orders__order_id__cancel_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/invoice": {
      "get": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Download Invoice",
        "description": "Generates and downloads official PDF Tax Invoice.",
        "operationId": "download_invoice_api_v1_orders__order_id__invoice_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/status": {
      "put": {
        "tags": [
          "5. Order System & Tracking"
        ],
        "summary": "Update Order Status",
        "description": "Updates order status (Staff/Admin only).",
        "operationId": "update_order_status_api_v1_orders__order_id__status_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          },
          {
            "name": "status_str",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Status Str"
            }
          },
          {
            "name": "tracking_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tracking Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": [
          "5. Orders Management"
        ],
        "summary": "Create Order",
        "operationId": "create_order_api_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__orders__order_router__CreateOrderPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "5. Orders Management"
        ],
        "summary": "Cancel Order",
        "operationId": "cancel_order_api_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{order_id}/invoice": {
      "get": {
        "tags": [
          "5. Orders Management"
        ],
        "summary": "Get Invoice Pdf",
        "operationId": "get_invoice_pdf_api_orders__order_id__invoice_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelOrderPayload": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "CancelOrderPayload"
      },
...
      "CheckoutPayload": {
        "properties": {
          "address_id": {
            "type": "integer",
            "title": "Address Id"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "default": "Razorpay"
          },
          "coupon_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coupon Code"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "address_id"
        ],
        "title": "CheckoutPayload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "quantity",
          "unit_price",
          "subtotal"
        ],
        "title": "OrderItemOut"
      },
      "OrderItemPayload": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "quantity"
        ],
        "title": "OrderItemPayload"
      },
      "OrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "order_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Number"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "discount_amount": {
            "type": "number",
            "title": "Discount Amount"
          },
          "net_amount": {
            "type": "number",
            "title": "Net Amount"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Number"
          },
          "shipping_address": {
            "type": "string",
            "title": "Shipping Address"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "total_amount",
          "discount_amount",
          "net_amount",
          "status",
          "payment_method",
          "payment_status",
          "shipping_address",
          "created_at"
        ],
        "title": "OrderOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__orders__order_router__CreateOrderPayload": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemPayload"
            },
            "type": "array",
            "title": "Items"
          },
          "shipping_address": {
            "type": "string",
            "title": "Shipping Address"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "default": "Razorpay"
          },
          "coupon_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coupon Code"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "items",
          "shipping_address"
        ],
        "title": "CreateOrderPayload"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}